requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django
django-crispy-forms==2.0
django-environ==0.10.0
django-js-asset==1.2.2
django-phonenumber-field==7.0.2
gunicorn==20.1.0
phonenumbers==8.12.29
Pillow==9.5.0
python-gettext==5.0
pytz==2021.1
sqlparse==0.4.1
tzdata==2023.3
whitenoise==5.3.0
